rerouted code in app js to peoplejs
