+++ data/people/hVztGXURs
@@ -1,0 +1,1 @@
+{"name":"Uncle bob","age":100,"favoriteColor":"red","_id":"hVztGXURs"}
