+++ server.js
@@ -1,13 +1,40 @@
-const app = require('./lib/app');
-const http = rquire('http');
 
-http
-  .createServer(app)
-  .listen(7890, () => console.log('Listeing on 7890')) {
-    bodyParser(req)
-      .then(body => {
-        people.creat({ name: body.name }, (err, data) => {
-          resizeBy.end();
-        })
-      })
-  }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+// const app = require('./lib/app');
+// const http = rquire('http');
+
+// http
+//   .createServer(app)
+//   .listen(7890, () => console.log('Listeing on 7890')) {
+//     bodyParser(req)
+//       .then(body => {
+//         people.creat({ name: body.name }, (err, data) => {
+//           resizeBy.end();
+//         })
+//       })
+//   }
