--- testData/store/R0ssoye4ib
@@ -1,1 +1,0 @@
-{"name":"ryan","_id":"R0ssoye4ib"}
